Have user confirm before they delete a course.

+++ public/js/coursesScript.js
@@ -7,11 +7,14 @@ delBtns.forEach((btn) => {
   btn.addEventListener("click", async () => {
     console.log(btn.id);
     try {
-      const response = await api.delete(`/delete/${btn.id}`);
-      console.log("Deleted:", response.data);
-      const row = btn.closest("tr");
-      if (row) {
-        row.remove();
+      const reply = confirm("Are you sure you want to delete this course?");
+      if (reply) {
+        const response = await api.delete(`/delete/${btn.id}`);
+        console.log("Deleted:", response.data);
+        const row = btn.closest("tr");
+        if (row) {
+          row.remove();
+        }
       }
     } catch (error) {
       console.log(error);
